RAFT: Make query work again over network. Move worker.go functions into respective files.
