+++ worker/predicate.go
@@ -17,8 +17,10 @@
 package worker
 
 import (
+	"bytes"
 	"context"
 	"io"
+	"sort"
 
 	"github.com/dgraph-io/dgraph/posting/types"
 	"github.com/dgraph-io/dgraph/task"
@@ -172,4 +174,62 @@ func (s *State) PopulateShard(ctx context.Context, pl *pool, group uint64) (int,
 	}
 	x.Trace(ctx, "Streaming complete for group: %v", group)
 	return count, nil
+}
+
+// PredicateData can be used to return data corresponding to a predicate over
+// a stream.
+func (w *grpcWorker) PredicateData(query *Payload, stream Worker_PredicateDataServer) error {
+	var group task.GroupKeys
+	uo := flatbuffers.GetUOffsetT(query.Data)
+	group.Init(query.Data, uo)
+	_ = group.Groupid()
+
+	// TODO(pawan) - Shift to CheckPoints once we figure out how to add them to the
+	// RocksDB library we are using.
+	// http://rocksdb.org/blog/2609/use-checkpoints-for-efficient-snapshots/
+	it := ws.dataStore.NewIterator()
+	defer it.Close()
+
+	for it.SeekToFirst(); it.Valid(); it.Next() {
+		k, v := it.Key(), it.Value()
+		pl := types.GetRootAsPostingList(v.Data(), 0)
+
+		// TODO: Check that key is part of the specified group id.
+		i := sort.Search(group.KeysLength(), func(i int) bool {
+			var t task.KC
+			x.Assertf(group.Keys(&t, i), "Unable to parse task.KC")
+			return bytes.Compare(k.Data(), t.KeyBytes()) <= 0
+		})
+
+		if i < group.KeysLength() {
+			// Found a match.
+			var t task.KC
+			x.Assertf(group.Keys(&t, i), "Unable to parse task.KC")
+
+			if bytes.Equal(k.Data(), t.KeyBytes()) && bytes.Equal(pl.Checksum(), t.ChecksumBytes()) {
+				// No need to send this.
+				continue
+			}
+		}
+
+		b := flatbuffers.NewBuilder(0)
+		bko := b.CreateByteVector(k.Data())
+		bvo := b.CreateByteVector(v.Data())
+		task.KVStart(b)
+		task.KVAddKey(b, bko)
+		task.KVAddVal(b, bvo)
+		kvoffset := task.KVEnd(b)
+		b.Finish(kvoffset)
+
+		p := Payload{Data: b.Bytes[b.Head():]}
+		if err := stream.Send(&p); err != nil {
+			return err
+		}
+		k.Free()
+		v.Free()
+	}
+	if err := it.Err(); err != nil {
+		return err
+	}
+	return nil
 }
