+++ worker/task.go
@@ -28,7 +28,6 @@ import (
 )
 
 const (
-	_xid_ = "_xid_"
 	_uid_ = "_uid_"
 )
 
@@ -36,46 +35,39 @@ const (
 // the instance which stores posting list corresponding to the predicate in the
 // query.
 func ProcessTaskOverNetwork(ctx context.Context, qu []byte) (result []byte, rerr error) {
-	q := new(task.Query)
-	x.ParseTaskQuery(q, qu)
-
+	q := task.GetRootAsQuery(qu, 0)
 	attr := string(q.Attr())
 	gid := BelongsTo(attr)
 	runHere := groups().ServesGroup(gid)
-
 	x.Trace(ctx, "runHere: %v attr: %v groupId: %v", runHere, attr, gid)
 
 	if runHere {
-		// No need for a network call, as this should be run from within
-		// this instance.
+		// No need for a network call, as this should be run from within this instance.
 		return processTask(qu)
 	}
 
-	// Using a worker client for the instance idx, we get the result of the query.
-	// TODO: Make this work again.
-	return
-	/*
-		pool := ws.GetPool(int(idx))
-		addr := pool.Addr
-		query := new(Payload)
-		query.Data = qu
+	// This needs to be send over the network.
+	// TODO: Send the request to multiple servers as described in Jeff Dean's talk.
+	addr := groups().AnyServer(gid)
+	pl := pools().get(addr)
 
-		conn, err := pool.Get()
-		if err != nil {
-			return []byte(""), err
-		}
-		defer pool.Put(conn)
-		c := NewWorkerClient(conn)
-		reply, err := c.ServeTask(context.Background(), query)
-		if err != nil {
-			x.TraceError(ctx, x.Wrapf(err, "Error while calling Worker.ServeTask"))
-			return []byte(""), err
-		}
+	conn, err := pl.Get()
+	if err != nil {
+		return result, x.Wrapf(err, "ProcessTaskOverNetwork: while retrieving connection.")
+	}
+	defer pl.Put(conn)
+	x.Trace(ctx, "Sending request to %v", addr)
 
-		x.Trace(ctx, "Reply from server. length: %v Addr: %v Attr: %v",
-			len(reply.Data), addr, attr)
-		return reply.Data, nil
-	*/
+	c := NewWorkerClient(conn)
+	reply, err := c.ServeTask(ctx, &Payload{Data: qu})
+	if err != nil {
+		x.TraceError(ctx, x.Wrapf(err, "Error while calling Worker.ServeTask"))
+		return []byte(""), err
+	}
+
+	x.Trace(ctx, "Reply from server. length: %v Addr: %v Attr: %v",
+		len(reply.Data), addr, attr)
+	return reply.Data, nil
 }
 
 // processTask processes the query, accumulates and returns the result.
@@ -171,21 +163,30 @@ func processTask(query []byte) ([]byte, error) {
 
 // ServeTask is used to respond to a query.
 func (w *grpcWorker) ServeTask(ctx context.Context, query *Payload) (*Payload, error) {
-	q := new(task.Query)
-	x.ParseTaskQuery(q, query.Data)
+	if ctx.Err() != nil {
+		return &Payload{}, ctx.Err()
+	}
+
+	q := task.GetRootAsQuery(query.Data, 0)
 	gid := BelongsTo(string(q.Attr()))
 	x.Trace(ctx, "Attribute: %q NumUids: %v groupId: %v ServeTask", q.Attr(), q.UidsLength(), gid)
 
 	reply := new(Payload)
-	var rerr error
-	// Request for xid <-> uid conversion should be handled by instance 0 and all
-	// other requests should be handled according to modulo sharding of the predicate.
-
-	// TODO: This should be done based on group rules, and not a modulo of num groups.
-	if groups().ServesGroup(gid) {
-		reply.Data, rerr = processTask(query.Data)
-	} else {
+	if !groups().ServesGroup(gid) {
 		log.Fatalf("attr: %q groupId: %v Request sent to wrong server.", q.Attr(), gid)
 	}
-	return reply, rerr
+
+	c := make(chan error, 1)
+	go func() {
+		var err error
+		reply.Data, err = processTask(query.Data)
+		c <- err
+	}()
+
+	select {
+	case <-ctx.Done():
+		return reply, ctx.Err()
+	case err := <-c:
+		return reply, err
+	}
 }
